Move database controller views (#2472)
* Move the database controller

* Move the database view

* Remove fields that cause errors in view rendering
